OD-17332 Bugsnag: Fixed typeerror - Cannot read properties of undefined (reading 'replace')

+++ client-html/src/js/containers/tags/containers/TagsFormBase.tsx
@@ -294,11 +294,13 @@ export class TagsFormBase extends React.PureComponent<FormProps, FormState> {
     const onConfirm = () => {
       const clone = JSON.parse(JSON.stringify(values));
 
-      const removePath = getDeepValue(clone, item.parent.replace(/\[[0-9]+]$/, ""));
+      if (item.parent) {
+        const removePath = getDeepValue(clone, item.parent.replace(/\[[0-9]+]$/, ""));
 
-      if (removePath) {
-        const deleteItem = item.parent.match(/\[(\d)]$/);
-        if (deleteItem && deleteItem.length > 0) removePath.splice(Number(deleteItem[1]), 1);
+        if (removePath) {
+          const deleteItem = item.parent.match(/\[(\d)]$/);
+          if (deleteItem && deleteItem.length > 0) removePath.splice(Number(deleteItem[1]), 1);
+        }
       }
 
       dispatch(change(TAGS_FORM_NAME, "childTags", clone.childTags));
